Journal system: Fix exception formatting
git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@2977 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/journal_exceptions.cc
@@ -44,9 +44,9 @@ throw ()
 	text  = "Journal file ";
 	text += _filename;
 	text += " contains record with type ";
-	text += toString(_code);
+	text += toString(static_cast<int>(_code));
 	text += " instead of the expected type ";
-	text += toString(_expectedcode);
+	text += toString(static_cast<int>(_expectedcode));
 }
 
 BadEvent_error::BadEvent_error
